+++ src/controllers/groupController.js
@@ -1,43 +1,23 @@
 import { prisma } from '../config/prisma.js';
 import { catchHandler } from '../lib/catchHandler.js';
 
-export const goodLikeCount = catchHandler(async (req, res) => {
+export const getGroupDetail = catchHandler(async (req, res) => {
   const { id } = req.params;
-
   const group = await prisma.group.findUnique({
     where: { id },
+    include: {
+      members: {
+        select: {
+          id: true,
+          nickName: true,
+          createdAt: true,
+          updatedAt: true,
+        },
+      },
+      groupBadge: true,
+      ownerPassword: false,
+    },
   });
 
-  if (!group) {
-    return res.status(400).send({ message: '그룹을 찾을 수 없습니다.' });
-  }
-
-  const updateGroup = await prisma.group.update({
-    where: { id },
-    data: { likeCount: (group.likeCount += 1) },
-  });
-
-  res.status(200).send(updateGroup);
-});
-
-export const badLikeCount = catchHandler(async (req, res) => {
-  const { id } = req.params;
-
-  const group = await prisma.group.findUnique({
-    where: { id },
-  });
-  if (!group) {
-    return res.status(400).send({ message: '그룹을 찾을 수 없습니다.' });
-  }
-
-  if (group.likeCount === 0) {
-    return res.status(400).send({ message: '취소할 추천 수가 없습니다.' });
-  }
-
-  const updateGroup = await prisma.group.update({
-    where: { id },
-    data: { likeCount: (group.likeCount -= 1) },
-  });
-
-  res.status(200).send(updateGroup);
+  res.status(200).send(group);
 });
